Extract abstract base class, which could be part of the Felix HTTP Jetty12 bundle. This abstracts away all the ugly code currently required to make websockets work.
